+++ iOS/TopCariving/TopCariving/View/OptionCardCell.swift
@@ -136,11 +136,3 @@ class OptionCardCell: UICollectionViewCell {
         addButton.setTitleColor(.hyundaiBlue, for: .normal)
     }
 }
-
-extension String {
-    static func decimalStyle(from number: Int) -> String {
-        let numberFormatter = NumberFormatter()
-        numberFormatter.numberStyle = .decimal
-        return numberFormatter.string(from: NSNumber(value: number)) ?? ""
-    }
-}
